backend/requirements.txt
flask==3.1.*
groq==0.24.*
piper-tts==1.2.*
pymongo==4.11.*
python-dotenv==1.1.*
gunicorn==23.0.*
